:star: product view

+++ models/offer.py
@@ -4,8 +4,8 @@ class Offer(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     product_id = db.Column(db.Integer, db.ForeignKey('product.id'), nullable=False)
     provider_id = db.Column(db.Integer, db.ForeignKey('provider.id'), nullable=False)
-    price = db.Column(db.Float, nullable=True)  # Precio
-    sale_price = db.Column(db.Float, nullable=True)  # Precio con descuento
+    price = db.Column(db.Integer, nullable=True)  # Precio
+    sale_price = db.Column(db.Integer, nullable=True)  # Precio con descuento
     pack_price = db.Column(db.Text, nullable=True)  # Precio con promocion, texto
     comment = db.Column(db.Text, unique=False, nullable=True)
     source = db.Column(db.Text, unique=False, nullable=True)
